Move Diff.Options to Options.Create
