+++ tools/diff-cli/core/api/core.api
@@ -1,54 +1,26 @@
 public final class io/matthewnelson/diff/core/Diff {
 	public static final field Companion Lio/matthewnelson/diff/core/Diff$Companion;
-	public static final fun apply (Ljava/lang/String;Ljava/lang/String;)V
-	public static final fun apply (Lokio/Path;Lokio/Path;)V
+	public static final fun apply (Ljava/lang/String;Ljava/lang/String;Z)V
+	public static final fun apply (Lokio/Path;Lokio/Path;Z)V
 	public static final fun create (Ljava/lang/String;Ljava/lang/String;Ljava/lang/String;)Ljava/lang/String;
-	public static final fun create (Ljava/lang/String;Ljava/lang/String;Ljava/lang/String;Lio/matthewnelson/diff/core/Diff$Options;)Ljava/lang/String;
+	public static final fun create (Ljava/lang/String;Ljava/lang/String;Ljava/lang/String;Lio/matthewnelson/diff/core/Options$Create;)Ljava/lang/String;
 	public static final fun create (Lokio/Path;Lokio/Path;Lokio/Path;)Lokio/Path;
-	public static final fun create (Lokio/Path;Lokio/Path;Lokio/Path;Lio/matthewnelson/diff/core/Diff$Options;)Lokio/Path;
+	public static final fun create (Lokio/Path;Lokio/Path;Lokio/Path;Lio/matthewnelson/diff/core/Options$Create;)Lokio/Path;
 	public static final fun readHeader (Ljava/lang/String;)Lio/matthewnelson/diff/core/Header;
 	public static final fun readHeader (Lokio/Path;)Lio/matthewnelson/diff/core/Header;
 }
 
 public final class io/matthewnelson/diff/core/Diff$Companion {
-	public final fun apply (Ljava/lang/String;Ljava/lang/String;)V
-	public final fun apply (Lokio/Path;Lokio/Path;)V
+	public final fun apply (Ljava/lang/String;Ljava/lang/String;Z)V
+	public final fun apply (Lokio/Path;Lokio/Path;Z)V
 	public final fun create (Ljava/lang/String;Ljava/lang/String;Ljava/lang/String;)Ljava/lang/String;
-	public final fun create (Ljava/lang/String;Ljava/lang/String;Ljava/lang/String;Lio/matthewnelson/diff/core/Diff$Options;)Ljava/lang/String;
+	public final fun create (Ljava/lang/String;Ljava/lang/String;Ljava/lang/String;Lio/matthewnelson/diff/core/Options$Create;)Ljava/lang/String;
 	public final fun create (Lokio/Path;Lokio/Path;Lokio/Path;)Lokio/Path;
-	public final fun create (Lokio/Path;Lokio/Path;Lokio/Path;Lio/matthewnelson/diff/core/Diff$Options;)Lokio/Path;
-	public static synthetic fun create$default (Lio/matthewnelson/diff/core/Diff$Companion;Ljava/lang/String;Ljava/lang/String;Ljava/lang/String;Lio/matthewnelson/diff/core/Diff$Options;ILjava/lang/Object;)Ljava/lang/String;
-	public static synthetic fun create$default (Lio/matthewnelson/diff/core/Diff$Companion;Lokio/Path;Lokio/Path;Lokio/Path;Lio/matthewnelson/diff/core/Diff$Options;ILjava/lang/Object;)Lokio/Path;
+	public final fun create (Lokio/Path;Lokio/Path;Lokio/Path;Lio/matthewnelson/diff/core/Options$Create;)Lokio/Path;
+	public static synthetic fun create$default (Lio/matthewnelson/diff/core/Diff$Companion;Ljava/lang/String;Ljava/lang/String;Ljava/lang/String;Lio/matthewnelson/diff/core/Options$Create;ILjava/lang/Object;)Ljava/lang/String;
+	public static synthetic fun create$default (Lio/matthewnelson/diff/core/Diff$Companion;Lokio/Path;Lokio/Path;Lokio/Path;Lio/matthewnelson/diff/core/Options$Create;ILjava/lang/Object;)Lokio/Path;
 	public final fun readHeader (Ljava/lang/String;)Lio/matthewnelson/diff/core/Header;
 	public final fun readHeader (Lokio/Path;)Lio/matthewnelson/diff/core/Header;
-}
-
-public final class io/matthewnelson/diff/core/Diff$Options {
-	public static final field Companion Lio/matthewnelson/diff/core/Diff$Options$Companion;
-	public static final field DEFAULT_EXT_NAME Ljava/lang/String;
-	public static final field STATIC_TIME Ljava/lang/String;
-	public final field diffFileExtensionName Ljava/lang/String;
-	public final field schema Lio/matthewnelson/diff/core/Diff$Schema;
-	public final field useStaticTime Z
-	public fun <init> ()V
-	public fun <init> (Lio/matthewnelson/diff/core/Diff$Options$Builder;)V
-	public fun <init> (Lkotlin/jvm/functions/Function1;)V
-	public fun equals (Ljava/lang/Object;)Z
-	public fun hashCode ()I
-	public fun toString ()Ljava/lang/String;
-}
-
-public final class io/matthewnelson/diff/core/Diff$Options$Builder {
-	public field schema Lio/matthewnelson/diff/core/Diff$Schema;
-	public field useStaticTime Z
-	public fun <init> ()V
-	public final fun diffFileExtensionName ()Ljava/lang/String;
-	public final fun diffFileExtensionName (Ljava/lang/String;)Lio/matthewnelson/diff/core/Diff$Options$Builder;
-	public final fun schema (Lio/matthewnelson/diff/core/Diff$Schema;)Lio/matthewnelson/diff/core/Diff$Options$Builder;
-	public final fun useStaticTime (Z)Lio/matthewnelson/diff/core/Diff$Options$Builder;
-}
-
-public final class io/matthewnelson/diff/core/Diff$Options$Companion {
 }
 
 public final class io/matthewnelson/diff/core/Diff$Schema : java/lang/Enum {
@@ -79,6 +51,37 @@ public class io/matthewnelson/diff/core/NoDiffException : java/lang/RuntimeExcep
 	public fun <init> (Ljava/lang/String;)V
 	public fun <init> (Ljava/lang/String;Ljava/lang/Throwable;)V
 	public final fun getMessage ()Ljava/lang/String;
+}
+
+public abstract class io/matthewnelson/diff/core/Options {
+}
+
+public final class io/matthewnelson/diff/core/Options$Create : io/matthewnelson/diff/core/Options {
+	public static final field Companion Lio/matthewnelson/diff/core/Options$Create$Companion;
+	public static final field DEFAULT_EXT_NAME Ljava/lang/String;
+	public static final field STATIC_TIME Ljava/lang/String;
+	public final field diffFileExtensionName Ljava/lang/String;
+	public final field schema Lio/matthewnelson/diff/core/Diff$Schema;
+	public final field useStaticTime Z
+	public fun <init> ()V
+	public fun <init> (Lio/matthewnelson/diff/core/Options$Create$Builder;)V
+	public fun <init> (Lkotlin/jvm/functions/Function1;)V
+	public fun equals (Ljava/lang/Object;)Z
+	public fun hashCode ()I
+	public fun toString ()Ljava/lang/String;
+}
+
+public final class io/matthewnelson/diff/core/Options$Create$Builder {
+	public field schema Lio/matthewnelson/diff/core/Diff$Schema;
+	public field useStaticTime Z
+	public fun <init> ()V
+	public final fun diffFileExtensionName ()Ljava/lang/String;
+	public final fun diffFileExtensionName (Ljava/lang/String;)Lio/matthewnelson/diff/core/Options$Create$Builder;
+	public final fun schema (Lio/matthewnelson/diff/core/Diff$Schema;)Lio/matthewnelson/diff/core/Options$Create$Builder;
+	public final fun useStaticTime (Z)Lio/matthewnelson/diff/core/Options$Create$Builder;
+}
+
+public final class io/matthewnelson/diff/core/Options$Create$Companion {
 }
 
 public abstract interface annotation class io/matthewnelson/diff/core/internal/InternalDiffApi : java/lang/annotation/Annotation {
